Library Updates
Removed Debug Statements from libraries so libraries are less memory intensive.

In ArrayList: Changed int tempCapacity to size_t tempCapacity to get rid of unsigned ints comparrison warning.

In Hashtable:
Removed 2nd Constructor... This was causing problems with this library, and 2 other libraries that depend on it.

in SDList:
Complete overhaul... this update fixed two issues

+++ lib/ArrayList/src/ArrayList.h
@@ -57,6 +57,7 @@ public:
         if (sizeType == DYNAMIC && loadFactor >= 0.8) {
             resize();
         }
+        // Add the item to the array
         if (count < arrayCapacity) {
             array[count++] = item;
         }
@@ -706,7 +707,7 @@ private:
         memcpy(newArray, array, count * sizeof(T));
         delete[] array;
         array = newArray;
-        int tempCapacity = newCapacity;
+        size_t tempCapacity = newCapacity;
         int oldCapacity = arrayCapacity;
         arrayCapacity = newCapacity;
         if(tempCapacity != arrayCapacity){
